EVM: Reorganize to keep internals internal
The next step will be to move some shared components into a new crate,
but keeping internals private is the first step.
