+++ actors/evm/src/interpreter/bytecode.rs
@@ -2,7 +2,7 @@
 
 use std::ops::Deref;
 
-use super::execution::opcodes;
+use super::opcodes;
 
 #[derive(Clone, Debug)]
 pub struct Bytecode {
