+++ actors/evm/src/interpreter/mod.rs
@@ -1,31 +1,29 @@
 pub mod address;
 pub mod bytecode;
-pub mod execution;
-pub mod instructions;
-pub mod memory;
-pub mod output;
-pub mod precompiles;
-pub mod stack;
-pub mod system;
-pub mod uints;
+mod execution;
+mod instructions;
+mod memory;
+mod output;
+mod precompiles;
+mod stack;
+mod system;
+mod uints;
 
 #[cfg(test)]
 pub mod test_util;
 
 pub use {
     bytecode::Bytecode,
-    execution::{execute, ExecutionState},
-    output::Output,
+    execution::{execute, opcodes, ExecutionState},
+    output::{Outcome, Output},
     system::System,
     uints::{U256, U512},
 };
 
-#[macro_export]
-macro_rules! abort {
-  ($code:ident, $msg:literal $(, $ex:expr)*) => {
-      fvm_sdk::vm::abort(
-          fvm_shared::error::ExitCode::$code.value(),
-          Some(format!($msg, $($ex,)*).as_str()),
-      )
-  };
+/// The kind of call-like instruction.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum CallKind {
+    Call,
+    DelegateCall,
+    StaticCall,
 }
