+++ actors/evm/tests/create.rs
@@ -1,9 +1,7 @@
 mod asm;
 
+use evm::ext::eam;
 use evm::interpreter::address::EthAddress;
-use evm::interpreter::instructions::lifecycle::{
-    Create2Params, CreateParams, EamReturn, CREATE2_METHOD_NUM, CREATE_METHOD_NUM,
-};
 use fil_actor_evm as evm;
 use fil_actors_runtime::EAM_ACTOR_ADDR;
 use fvm_ipld_encoding::ipld_block::IpldBlock;
@@ -30,7 +28,7 @@ mstore # "bytecode"
 test_create:
     jumpdest
     push1 0x10   # in size (16 bytes)
-    push2 0x0110 # in offset 
+    push2 0x0110 # in offset
     push1 0x01   # value (attoFil)
     create
     %return_stack_word()
@@ -55,7 +53,7 @@ fn test_create() {
     let mut rt = util::construct_and_verify(contract);
 
     let fake_eth_addr = EthAddress(hex_literal::hex!("CAFEB0BA00000000000000000000000000000000"));
-    let fake_ret = EamReturn {
+    let fake_ret = eam::CreateReturn {
         actor_id: 12345,
         eth_address: fake_eth_addr,
         robust_address: Some((&fake_eth_addr).try_into().unwrap()),
@@ -64,12 +62,12 @@ fn test_create() {
     let salt =
         hex_literal::hex!("0000000000000000000000000000000000000000000000796573206D616E6921");
 
-    let create2_params = Create2Params {
+    let create2_params = eam::Create2Params {
         code: hex_literal::hex!("666F6F206261722062617A20626F7879").to_vec(),
         salt,
     };
 
-    let mut create_params = CreateParams {
+    let mut create_params = eam::CreateParams {
         code: hex_literal::hex!("666F6F206261722062617A20626F7879").to_vec(),
         nonce: 1,
     };
@@ -83,7 +81,7 @@ fn test_create() {
 
         rt.expect_send_simple(
             EAM_ACTOR_ADDR,
-            CREATE_METHOD_NUM,
+            eam::CREATE_METHOD_NUM,
             IpldBlock::serialize_cbor(&create_params).unwrap(),
             TokenAmount::from_atto(1),
             IpldBlock::serialize_cbor(&fake_ret).unwrap(),
@@ -105,7 +103,7 @@ fn test_create() {
 
         rt.expect_send_simple(
             EAM_ACTOR_ADDR,
-            CREATE_METHOD_NUM,
+            eam::CREATE_METHOD_NUM,
             IpldBlock::serialize_cbor(&create_params).unwrap(),
             TokenAmount::from_atto(1),
             IpldBlock::serialize_cbor(&fake_ret).unwrap(),
@@ -127,7 +125,7 @@ fn test_create() {
 
         rt.expect_send_simple(
             EAM_ACTOR_ADDR,
-            CREATE2_METHOD_NUM,
+            eam::CREATE2_METHOD_NUM,
             IpldBlock::serialize_cbor(&create2_params).unwrap(),
             TokenAmount::from_atto(1),
             IpldBlock::serialize_cbor(&fake_ret).unwrap(),
